+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Command `ship-mate-hide` bound to `C-c C-q` in
+  `ship-mate-dinghy-mode`. It hides a running compilation like other
+  hidden compilations.
+
 ## [0.3.2]
 
 ### Added